// sdk/node/src/generate/account/transfer/model_futures_account_transfer_in_resp.ts
// Code generated by Kucoin Universal SDK Generator; DO NOT EDIT.

import { instanceToPlain, Exclude, plainToInstance } from 'class-transformer';
import { RestResponse } from '@model/common';
import { Response } from '@internal/interfaces/response';
export class FuturesAccountTransferInResp
    implements Response<FuturesAccountTransferInResp, RestResponse>
{
    /**
     *
     */
    data?: string;
    /**
     * common response
     */
    @Exclude()
    commonResponse?: RestResponse;

    setCommonResponse(response: RestResponse): void {
        this.commonResponse = response;
    }

    fromJson(input: string): FuturesAccountTransferInResp {
        const jsonObject = JSON.parse(input);
        return plainToInstance(FuturesAccountTransferInResp, { data: jsonObject });
    }

    toJson(): string {
        return JSON.stringify(instanceToPlain(this.data));
    }

    fromObject(jsonObject: Object): FuturesAccountTransferInResp {
        return plainToInstance(FuturesAccountTransferInResp, { data: jsonObject });
    }
}